{DO NOT MERGE} Correctly create and surface data service in Flink runner
The data service was not previously being created correctly, and its
API service descriptor needed to be surfaced to the Flink operator.
The data service is now lazily instantiated by a JobResourceManager, and
surfaced to the Flink operator via the EnvironmentSession.

DO NOT MERGE until base branch flink-portable-runner is fixed and this
commit can be rebased.  It does not currently compile, preventing
checking compile errors.